Made it so that whenver a player joins they are added to a list for 10 seconds
That way, we can check if the player has just joined and prevent doing
stuff that would interfere with the onPlayerJoin event handelers!
Already implemented it in Tips so that they don't show up when the
player has just joined.
Also added some comments to a lot of the code!